make load_data robust to being run more than once on the same day
the "drop table" command was failing because there are dependent tables.  the
"create table" command was then failing because the table already existed, but
because we didn't have ON_ERROR_STOP set, it would continue with the rest of
the process and just insert more rows into the existing table

now the script is more robust as we use ON_ERROR_STOP throughout. this also
cleans up the output a bit.